src/requirements.txt
# python<=3.8.2
# pip install -r requirements.txt

# YOLOv5 ----------------------------------------

matplotlib>=3.2.2
numpy>=1.18.5
opencv-python==4.5.2.52
opencv-contrib-python==3.4.11.45
Pillow>=7.1.2
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
torch>=1.7.0
torchvision>=0.8.1
tqdm>=4.64.0
protobuf<=3.20.1

# plotting ------------------------------------

pandas>=1.1.4
seaborn>=0.11.0

# deep_sort -----------------------------------

easydict

# torchreid -----------------------------------

Cython
h5py
six
tb-nightly
future
yacs
gdown
flake8
yapf
isort==4.3.21
imageio

# django --------------------------------

django>=3.1.3
djangorestframework>=3.12.2
django-filter>=22.1
markdown>=3.4.1
django-cors-headers==3.13.0
pandas
Pillow
opencv-python

# django admin extensions --------------------

django-db-views
django-extensions
ipython